requirements.txt
python-socketio==5.12.1
python-escpos==3.1
pyusb==1.3.1
